% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCNEFE.R
\name{download_CNEFE}
\alias{download_CNEFE}
\title{Download CNEFE datasets}
\usage{
download_CNEFE(state = "all", dest_path = "", overwrite = FALSE)
}
\arguments{
\item{state}{state abbreviation of datasets to be downloaded; "all" downloads data from all states}

\item{dest_path}{destination directory of downloaded files (defaults to current working directory)}

\item{overwrite}{whether to overwrite previosly downloaded files (if present) or not}
}
\value{
local path of downloaded compressed files from IBGE
}
\description{
Downloads CNEFE datasets developed from IBGE - Brazilian Institute of Geography and Statistics. Data provided by IBGE consists of a countrywide georeferenced register of addresses, which classifies addresses by type (residential, non-residential, etc.).
}
\examples{
temp1 <- download_CNEFE(state="all",dest_path="",overwrite=FALSE)
temp2 <- download_CNEFE(state="CE",dest_path="",overwrite=TRUE)
}
