% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCNEFE.R
\name{get_CNEFE}
\alias{get_CNEFE}
\title{Download and process CNEFE datasets}
\usage{
get_CNEFE(state = "all", dest_path = "", output_type = "df")
}
\arguments{
\item{state}{state abbreviation of datasets to be downloaded; "all" downloads data from all states.}

\item{dest_path}{directory path of downloaded files (defaults to current working directory).}

\item{output_type}{"df" outputs data frame object(s). "sf" outputs a simple features object using geographical coordinated from adresses.}
}
\value{
If you are processing a single brazilian state, function outputs a data frame or sf object; If you are processing "all", output is a list of data frames or a list of sf objects.
}
\description{
Downloads and processes CNEFE datasets developed from IBGE - Brazilian Institute of Geography and Statistics. Data provided by IBGE consists of a countrywide georeferenced register of addresses, which classifies addresses by type (residential, non-residential, etc.).
After locally downloading files from IBGE server, function properly reads prevously downloaded text files from CNEFE, tidies data and treats geographical coordinates of addresses. In case of missing data, imputes coordinates from census tract.
}
\examples{
temp1 <- get_CNEFE(state="all",dest_path="",output_type="df")
temp2 <- get_CNEFE(state="CE",dest_path="",output_type="df")
}
